+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/RecordMetaDataBuilder.java
@@ -115,6 +115,8 @@ public class RecordMetaDataBuilder implements RecordMetaDataProvider {
     private RecordMetaData recordMetaData;
     @Nonnull
     private final Map<String, Descriptors.FileDescriptor> explicitDependencies;
+    private long subspaceKeyCounter = 0;
+    private boolean usesSubspaceKeyCounter = false;
 
     /**
      * Creates a blank builder.
@@ -293,9 +295,24 @@ public class RecordMetaDataBuilder implements RecordMetaDataProvider {
                                @Nonnull Descriptors.FileDescriptor[] dependencies,
                                boolean processExtensionOptions) {
         recordsDescriptor = buildFileDescriptor(metaDataProto.getRecords(), dependencies);
+        loadSubspaceKeySettingsFromProto(metaDataProto);
         unionDescriptor = initRecordTypes(recordsDescriptor, processExtensionOptions);
         loadProtoExceptRecords(metaDataProto);
         processSchemaOptions(processExtensionOptions);
+    }
+
+    private void loadSubspaceKeySettingsFromProto(RecordMetaDataProto.MetaData metaDataProto) {
+        if (metaDataProto.hasSubspaceKeyCounter() && !metaDataProto.getUsesSubspaceKeyCounter()) {
+            throw new MetaDataProtoDeserializationException(new MetaDataException("subspaceKeyCounter is set but usesSubspaceKeyCounter is not set in the meta-data proto"));
+        }
+        if (metaDataProto.getUsesSubspaceKeyCounter() && !metaDataProto.hasSubspaceKeyCounter()) {
+            throw new MetaDataProtoDeserializationException(new MetaDataException("usesSubspaceKeyCounter is set but subspaceKeyCounter is not set in the meta-data proto"));
+        }
+        if (!usesSubspaceKeyCounter()) {
+            // Only read from the proto if user has not explicitly enabled it already.
+            usesSubspaceKeyCounter = metaDataProto.getUsesSubspaceKeyCounter();
+        }
+        subspaceKeyCounter = Long.max(subspaceKeyCounter, metaDataProto.getSubspaceKeyCounter()); // User might have set the counter already.
     }
 
     private void loadFromFileDescriptor(@Nonnull Descriptors.FileDescriptor fileDescriptor,
@@ -898,6 +915,9 @@ public class RecordMetaDataBuilder implements RecordMetaDataProvider {
         if (index.getAddedVersion() <= 0) {
             index.setAddedVersion(index.getLastModifiedVersion());
         }
+        if (usesSubspaceKeyCounter && !index.hasExplicitSubspaceKey()) {
+            index.setSubspaceKey(++subspaceKeyCounter);
+        }
         indexes.put(index.getName(), index);
     }
 
@@ -1065,6 +1085,77 @@ public class RecordMetaDataBuilder implements RecordMetaDataProvider {
     }
 
     /**
+     * Enable counter-based subspace keys assignment.
+     *
+     * <p>
+     * If enabled, index subspace keys will be set using a counter instead of defaulting to the indexes' names. This
+     * must be called prior to setting the records descriptor (for example {@link #setRecords(Descriptors.FileDescriptor)}).
+     * </p>
+     *
+     * <p>
+     * Existing clients should be careful about enabling this feature. The name of an index is the default
+     * value of its subspace key when counter-based subspace keys are disabled. If a subspace key was not set
+     * explicitly before, enabling the counter-based scheme will change the index's subspace key. Note that
+     * it is important that the subspace key of an index that has data does not change.
+     * See {@link Index#setSubspaceKey(Object)} for more details.
+     * </p>
+     *
+     * @return this builder
+     */
+    @Nonnull
+    public RecordMetaDataBuilder enableCounterBasedSubspaceKeys() {
+        if (recordsDescriptor != null) {
+            throw new MetaDataException("Records descriptor has already been set.");
+        }
+        this.usesSubspaceKeyCounter = true;
+        return this;
+    }
+
+    /**
+     * Checks if counter-based subspace key assignment is used.
+     * @return {@code true} if the subspace key counter is used
+     */
+    public boolean usesSubspaceKeyCounter() {
+        return usesSubspaceKeyCounter;
+    }
+
+    /**
+     * Get the current value of the index subspace key counter. If it is not enabled, the value will be 0.
+     * @return the current value of the index subspace key counter
+     * @see #enableCounterBasedSubspaceKeys()
+     */
+    public long getSubspaceKeyCounter() {
+        return subspaceKeyCounter;
+    }
+
+    /**
+     * Set the initial value of the subspace key counter. This method can be handy when users want to assign
+     * subspace keys using a counter, but their indexes already have subspace keys that may conflict with the
+     * counter-based assignment.
+     *
+     * <p>
+     * Note that the new counter must be greater than the current value. Also, users must first enable this feature by
+     * calling {@link #enableCounterBasedSubspaceKeys()} before updating the counter value.
+     * </p>
+     *
+     * @param subspaceKeyCounter the new value
+     * @return this builder
+     * @see #enableCounterBasedSubspaceKeys()
+     */
+    @Nonnull
+    public RecordMetaDataBuilder setSubspaceKeyCounter(long subspaceKeyCounter) {
+        if (!usesSubspaceKeyCounter()) {
+            throw new MetaDataException("Counter-based subspace keys not enabled");
+        }
+        if (subspaceKeyCounter <= this.subspaceKeyCounter) {
+            throw new MetaDataException(String.format("Subspace key counter must be set to a value greater than its current value (%d)",
+                    this.subspaceKeyCounter));
+        }
+        this.subspaceKeyCounter = subspaceKeyCounter;
+        return this;
+    }
+
+    /**
      * If there is only one record type, get it.
      * @return the only type defined for this store.
      */
@@ -1159,9 +1250,9 @@ public class RecordMetaDataBuilder implements RecordMetaDataProvider {
     @Nonnull
     public RecordMetaData build(boolean validate) {
         Map<String, RecordType> recordTypeBuilders = new HashMap<>();
-        RecordMetaData metaData = new RecordMetaData(recordsDescriptor, getUnionDescriptor(), unionFields, recordTypeBuilders,
+        RecordMetaData metaData = new RecordMetaData(recordsDescriptor, unionDescriptor, unionFields, recordTypeBuilders,
                 indexes, universalIndexes, formerIndexes,
-                splitLongRecords, storeRecordVersions, version, recordCountKey);
+                splitLongRecords, storeRecordVersions, version, subspaceKeyCounter, usesSubspaceKeyCounter, recordCountKey);
         for (RecordTypeBuilder recordTypeBuilder : this.recordTypes.values()) {
             KeyExpression primaryKey = recordTypeBuilder.getPrimaryKey();
             if (primaryKey != null) {